Move common harmony patch stuff to its own file
Cleans up the startup logic a bit
